#35 [REFACTOR] SearchInput 컴포넌트 재사용 가능하게 수정
