+++ src/components/Header/Header.jsx
@@ -39,7 +39,7 @@ export default function Header() {
   return (
     <div className='relative'>
       {searchClick ? (
-        <div className='absolute z-50 duration-500'>
+        <div className='absolute z-50'>
           <SearchHeader
             searchClick={searchClick}
             setSearchClick={setSearchClick}
@@ -48,7 +48,7 @@ export default function Header() {
       ) : (
         ''
       )}
-      <div className='flex h-[130px] justify-center items-center duration-500'>
+      <div className='flex h-[130px] justify-center items-center'>
         <div className='flex w-[1280px] h-[130px] justify-center items-center border-b border-gray-scale-8'>
           <div className='flex flex-row w-[1280px] h-[76px] py-[27px] justify-between items-center'>
             <div className='flex w-[216px] h-[61px]'>
